CSRF-Schutz für weitere Addons (#1553)

+++ redaxo/src/addons/structure/plugins/content/pages/templates.php
@@ -19,21 +19,27 @@ $error = '';
 $content = '';
 $message = '';
 
+$csrfToken = rex_csrf_token::factory('structure_content_template');
+
 if ($function == 'delete') {
-    $del = rex_sql::factory();
-    $del->setQuery('SELECT ' . rex::getTablePrefix() . 'article.id,' . rex::getTablePrefix() . 'template.name FROM ' . rex::getTablePrefix() . 'article
+    if (!$csrfToken->isValid()) {
+        $error = rex_i18n::msg('csrf_token_invalid');
+    } else {
+        $del = rex_sql::factory();
+        $del->setQuery('SELECT ' . rex::getTablePrefix() . 'article.id,' . rex::getTablePrefix() . 'template.name FROM ' . rex::getTablePrefix() . 'article
         LEFT JOIN ' . rex::getTablePrefix() . 'template ON ' . rex::getTablePrefix() . 'article.template_id=' . rex::getTablePrefix() . 'template.id
         WHERE ' . rex::getTablePrefix() . 'article.template_id="' . $template_id . '" LIMIT 0,10');
 
-    if ($del->getRows() > 0 || rex_template::getDefaultId() == $template_id) {
-        $error = rex_i18n::msg('cant_delete_template_because_its_in_use', rex_i18n::msg('id') . ' = ' . $template_id);
-    } else {
-        $del->setQuery('DELETE FROM ' . rex::getTablePrefix() . 'template WHERE id = "' . $template_id . '" LIMIT 1'); // max. ein Datensatz darf loeschbar sein
-        rex_file::delete(rex_path::addonCache('templates', $template_id . '.template'));
-        $success = rex_i18n::msg('template_deleted');
-        $success = rex_extension::registerPoint(new rex_extension_point('TEMPLATE_DELETED', $success, [
-            'id' => $template_id,
-        ]));
+        if ($del->getRows() > 0 || rex_template::getDefaultId() == $template_id) {
+            $error = rex_i18n::msg('cant_delete_template_because_its_in_use', rex_i18n::msg('id') . ' = ' . $template_id);
+        } else {
+            $del->setQuery('DELETE FROM ' . rex::getTablePrefix() . 'template WHERE id = "' . $template_id . '" LIMIT 1'); // max. ein Datensatz darf loeschbar sein
+            rex_file::delete(rex_path::addonCache('templates', $template_id . '.template'));
+            $success = rex_i18n::msg('template_deleted');
+            $success = rex_extension::registerPoint(new rex_extension_point('TEMPLATE_DELETED', $success, [
+                'id' => $template_id,
+            ]));
+        }
     }
 } elseif ($function == 'edit') {
     $legend = rex_i18n::msg('edit_template') . ' <small class="rex-primary-id">' . rex_i18n::msg('id') . ' = ' . $template_id . '</small>';
@@ -58,7 +64,10 @@ if ($function == 'delete') {
 }
 
 if ($function == 'add' or $function == 'edit') {
-    if ($save == 'ja') {
+    if ($save == 'ja' && !$csrfToken->isValid()) {
+        echo rex_view::error(rex_i18n::msg('csrf_token_invalid'));
+        $save = 'nein';
+    } elseif ($save == 'ja') {
         $active = rex_post('active', 'int');
         $templatename = rex_post('templatename', 'string');
         $template = rex_post('content', 'string');
@@ -425,6 +434,7 @@ if ($function == 'add' or $function == 'edit') {
 
         $content = '
             <form id="rex-form-template" action="' . rex_url::currentBackendPage(['start' => rex_request('start', 'int')]) . '" method="post">
+                ' . $csrfToken->getHiddenField() . '
                 ' . $content . '
             </form>
 
@@ -503,7 +513,7 @@ if ($OUT) {
 
     $list->addColumn(rex_i18n::msg('template_delete'), '<i class="rex-icon rex-icon-delete"></i> ' . rex_i18n::msg('delete'));
     $list->setColumnLayout(rex_i18n::msg('template_delete'), ['', '<td class="rex-table-action">###VALUE###</td>']);
-    $list->setColumnParams(rex_i18n::msg('template_delete'), ['function' => 'delete', 'template_id' => '###id###']);
+    $list->setColumnParams(rex_i18n::msg('template_delete'), ['function' => 'delete', 'template_id' => '###id###'] + $csrfToken->getUrlParams());
     $list->addLinkAttribute(rex_i18n::msg('template_delete'), 'data-confirm', rex_i18n::msg('confirm_delete_template'));
 
     $list->setNoRowsMessage(rex_i18n::msg('templates_not_found'));
